Added basic uploading and downloading of translations
Summary:
This commit adds basic support and infrastructure for uploading messages to Tolgee and downloading messages from Tolgee and writing them to local files. This is based on the import and export functionality in Tolgee.

This is a somewhat rough implementation which still needs polishing. It should be considered a functional prototype.

Reviewers: siepkes

Reviewed By: siepkes

Differential Revision: https://phabricator.serviceplanet.nl/D5101